+++ module-apps/application-calendar/windows/CalendarEventsOptionsWindow.hpp
@@ -1,0 +1,16 @@
+#pragma once
+#include "OptionWindow.hpp"
+
+namespace gui
+{
+    class CalendarEventsOptions : public OptionWindow
+    {
+      public:
+        CalendarEventsOptions(app::Application *app);
+        ~CalendarEventsOptions() override = default;
+
+      private:
+        auto eventsOptionsList() -> std::list<gui::Option>;
+        auto eventDelete() -> bool;
+    };
+}; // namespace gui